refactor(desktop): extract acceptance support
