+++ apps/desktop/scripts/support/packaged-telegram/acceptance-deadline.ts
