+++ apps/desktop/scripts/support/packaged-telegram/daemon-utility.ts
