+++ apps/desktop/scripts/support/packaged-telegram/disposable-keychain.ts
