+++ apps/desktop/scripts/support/packaged-telegram/main-entry.ts
