+++ apps/desktop/scripts/support/packaged-telegram/package-acceptance.d.mts
