+++ apps/desktop/scripts/support/packaged-telegram/package-acceptance.mjs
