+++ apps/desktop/scripts/support/packaged-telegram/process-safety.ts
