+++ apps/desktop/scripts/support/packaged-telegram/startup-mode.ts
